Day 03 (#3)
* Add input files
* Add brute force part 1 and fix warnings
* Rough part 2 day 3

+++ day_02/test.cpp
@@ -12,7 +12,7 @@ struct PasswordConfig {
 
 bool is_valid_password(const PasswordConfig &passwordConfig) {
   auto &password = passwordConfig.password;
-  size_t count =
+  int count =
       std::count(password.begin(), password.end(), passwordConfig.character);
   return (count >= passwordConfig.min && count <= passwordConfig.max);
 }
